requirements.txt
onnxruntime-qnn
numpy
pillow
tokenizers
